Add type tests
