+++ src/createQueryClient.spec.ts
@@ -511,42 +511,3 @@ describe('options', () => {
     expect(result.executed).toBe(true)
   })
 })
-
-describe('setQueryData', () => {
-  test('updates the query data', async () => {
-    const { setQueryData } = createQueryClient()
-    const untypedTag = tag()
-    const numberTag = tag<number>()
-    const stringTag = tag<string>()
-
-    const action = (arg: number) => 'foo'
-
-    setQueryData(action, (data) => {
-      return 'hello world'
-    })
-
-    setQueryData(action, [1], (data) => {
-      return 'hello world'
-    })
-
-    setQueryData(stringTag, (data) => {
-      return 'hello world'
-    })
-
-    setQueryData(numberTag, (data) => {
-      return 2
-    })
-
-    setQueryData(untypedTag, (data) => {
-      return 'hello world'
-    })
-
-    setQueryData([stringTag, numberTag], (data) => {
-      return 'hello world'
-    })
-
-    setQueryData([stringTag, numberTag, untypedTag], (data) => {
-      return 'hello world'
-    })
-  })
-})
